+++ src/main/java/dandi/dandi/post/adapter/in/web/dto/TemperatureRequest.java
@@ -1,4 +1,4 @@
-package dandi.dandi.post.web.in;
+package dandi.dandi.post.adapter.in.web.dto;
 
 public class TemperatureRequest {
