+++ src/test/java/dandi/dandi/common/AcceptanceTest.java
@@ -19,10 +19,10 @@ import dandi.dandi.auth.adapter.out.jwt.RefreshTokenManagerAdapter;
 import dandi.dandi.auth.application.port.in.TokenResponse;
 import dandi.dandi.auth.application.port.out.oauth.OAuthClientPort;
 import dandi.dandi.config.AsyncTestConfig;
+import dandi.dandi.post.adapter.in.web.dto.OutfitFeelingRequest;
+import dandi.dandi.post.adapter.in.web.dto.PostRegisterRequest;
+import dandi.dandi.post.adapter.in.web.dto.TemperatureRequest;
 import dandi.dandi.post.application.port.in.PostRegisterResponse;
-import dandi.dandi.post.web.in.OutfitFeelingRequest;
-import dandi.dandi.post.web.in.PostRegisterRequest;
-import dandi.dandi.post.web.in.TemperatureRequest;
 import io.restassured.RestAssured;
 import java.time.Clock;
 import org.junit.jupiter.api.AfterEach;
