+++ src/test/java/dandi/dandi/post/acceptance/PostAcceptanceTest.java
@@ -25,6 +25,9 @@ import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import dandi.dandi.common.AcceptanceTest;
+import dandi.dandi.post.adapter.in.web.dto.OutfitFeelingRequest;
+import dandi.dandi.post.adapter.in.web.dto.PostRegisterRequest;
+import dandi.dandi.post.adapter.in.web.dto.TemperatureRequest;
 import dandi.dandi.post.application.port.in.FeedResponse;
 import dandi.dandi.post.application.port.in.LikedPostResponse;
 import dandi.dandi.post.application.port.in.LikedPostResponses;
@@ -36,9 +39,6 @@ import dandi.dandi.post.application.port.in.PostImageRegisterResponse;
 import dandi.dandi.post.application.port.in.PostRegisterResponse;
 import dandi.dandi.post.application.port.in.PostResponse;
 import dandi.dandi.post.application.port.in.PostWriterResponse;
-import dandi.dandi.post.web.in.OutfitFeelingRequest;
-import dandi.dandi.post.web.in.PostRegisterRequest;
-import dandi.dandi.post.web.in.TemperatureRequest;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.io.File;
